% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plan.R
\name{add_op}
\alias{add_op}
\title{Append an operation to a plan}
\usage{
add_op(plan, node)
}
\arguments{
\item{plan}{A \code{gds_plan}}

\item{node}{Operation node (list)}
}
\value{
Updated plan
}
\description{
Append an operation to a plan
}
