% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gds-class.R
\name{add_provenance_node}
\alias{add_provenance_node}
\title{Append a provenance node to metadata}
\usage{
add_provenance_node(metadata, op_name, params, inputs = list())
}
\arguments{
\item{metadata}{Metadata list from \code{\link[=gds_metadata]{gds_metadata()}}}

\item{op_name}{Operation name}

\item{params}{Named list of parameters}

\item{inputs}{Parent node identifiers}
}
\value{
Updated metadata list
}
\description{
Append a provenance node to metadata
}
