% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verb-align.R
\name{align}
\alias{align}
\title{Align subjects into a consensus space}
\usage{
align(x, family)
}
\arguments{
\item{x}{A plan, source, or realised GDS}

\item{family}{A \code{\link{MapFamily}} object describing per-subject transforms}
}
\value{
Updated plan
}
\description{
Align subjects into a consensus space
}
