% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verb-align.R
\name{align_eager}
\alias{align_eager}
\title{Eagerly align subjects and compute immediately}
\usage{
align_eager(x, ...)
}
\arguments{
\item{x}{Plan, source, or realized GDS}

\item{...}{Arguments passed to align(), then compute()}
}
\value{
A realized GDS object
}
\description{
Eagerly align subjects and compute immediately
}
