% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/space-utils.R
\name{apply_common_mask}
\alias{apply_common_mask}
\title{Apply a common mask to two realised GDS objects}
\usage{
apply_common_mask(g1, g2, rule = c("intersection", "union"))
}
\arguments{
\item{g1}{First GDS}

\item{g2}{Second GDS}

\item{rule}{Mask rule ("intersection" or "union")}
}
\value{
A list with subsetted \code{g1}, \code{g2}, and indices \code{idx1}, \code{idx2}
}
\description{
Apply a common mask to two realised GDS objects
}
