% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gds-class.R
\name{assay}
\alias{assay}
\title{Extract a single assay from a GDS object}
\usage{
assay(x, name = "beta", ...)
}
\arguments{
\item{x}{A GDS object}

\item{name}{Assay name (default: "beta")}

\item{...}{Additional arguments}
}
\value{
A 3D array
}
\description{
Extract a single assay from a GDS object
}
