% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assay-registry.R
\name{assay_info}
\alias{assay_info}
\title{Retrieve assay metadata}
\usage{
assay_info(name)
}
\arguments{
\item{name}{Assay name}
}
\value{
Assay metadata list or NULL if not found
}
\description{
Retrieve assay metadata
}
