% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gds-class.R
\name{assays}
\alias{assays}
\title{Extract assays from a GDS object}
\usage{
assays(x)
}
\arguments{
\item{x}{A GDS object}
}
\value{
Named list of 3D arrays
}
\description{
Extract assays from a GDS object
}
