% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compat.R
\name{assert_compatible_spaces}
\alias{assert_compatible_spaces}
\title{Assert compatibility between two spaces}
\usage{
assert_compatible_spaces(g1, g2, fields = c("type", "dim", "template_id"))
}
\arguments{
\item{g1}{First GDS or space}

\item{g2}{Second GDS or space}

\item{fields}{Fields to compare (subset of c("type","dim","template_id"))}
}
\value{
TRUE if compatible; otherwise errors
}
\description{
Assert compatibility between two spaces
}
\examples{
sp1 <- space_voxel(c(2,2,2), diag(4), storage = "dense")
sp2 <- space_voxel(c(2,2,2), diag(4), storage = "dense")
assert_compatible_spaces(sp1, sp2)
}
