% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assay-registry.R
\name{can_map_linear}
\alias{can_map_linear}
\title{Test whether an assay can be linearly mapped}
\usage{
can_map_linear(name)
}
\arguments{
\item{name}{Assay name}
}
\value{
Logical, TRUE if assay can be linearly mapped
}
\description{
Test whether an assay can be linearly mapped
}
