% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.R
\name{canonicalize_node}
\alias{canonicalize_node}
\title{Canonicalize an operation node}
\usage{
canonicalize_node(node)
}
\arguments{
\item{node}{Operation node list}
}
\value{
Canonicalized node list
}
\description{
Canonicalize an operation node
}
