% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gds-class.R
\name{col_data}
\alias{col_data}
\title{Extract column (subject) metadata from a GDS object}
\usage{
col_data(x)
}
\arguments{
\item{x}{A GDS object}
}
\value{
Data frame with subject-level metadata
}
\description{
Extract column (subject) metadata from a GDS object
}
