% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/space-utils.R
\name{common_mask}
\alias{common_mask}
\title{Compute a common mask between two spaces}
\usage{
common_mask(g1, g2, rule = c("intersection", "union"))
}
\arguments{
\item{g1}{First GDS or space}

\item{g2}{Second GDS or space}

\item{rule}{Either "intersection" or "union"}
}
\value{
A list with integer indices \code{idx1} and \code{idx2} for subsetting each
}
\description{
Compute a common mask between two spaces
}
\examples{
sp1 <- space_voxel(c(2,2,1), diag(4), mask_bitmap = array(c(TRUE,FALSE,TRUE,TRUE), c(2,2,1)), storage = "packed")
sp2 <- space_voxel(c(2,2,1), diag(4), mask_bitmap = array(c(TRUE,TRUE,FALSE,TRUE), c(2,2,1)), storage = "packed")
common_mask(sp1, sp2, rule = "intersection")
}
