% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.R
\name{compute}
\alias{compute}
\title{Materialise a plan into a realised GDS}
\usage{
compute(
  x,
  sink = c("memory", "h5"),
  path = NULL,
  sink_options = list(),
  block = NULL,
  assays = NULL,
  ...
)
}
\arguments{
\item{x}{Plan or source}

\item{sink}{Storage sink ("memory" for in-memory data, "h5" to write a GDS HDF5 store)}

\item{path}{Output path when \code{sink = "h5"}}

\item{sink_options}{Optional list of sink-specific options}

\item{block}{Optional list specifying block indices to request from the adapter}

\item{assays}{Optional character vector to return raw arrays instead of GDS object}

\item{...}{Reserved for future use}
}
\value{
A realised \code{\link{gds}} object
}
\description{
Materialise a plan into a realised GDS
}
