% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gds-class.R
\name{contrast_data}
\alias{contrast_data}
\title{Extract contrast-level metadata from a GDS object}
\usage{
contrast_data(x)
}
\arguments{
\item{x}{A GDS object or plan}
}
\value{
Data frame with one row per contrast/repeated-measure level
}
\description{
Extract contrast-level metadata from a GDS object
}
