% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive.R
\name{derive}
\alias{derive}
\title{Record a derivation operation in a plan}
\usage{
derive(x, what = c("var", "se", "t", "z"), options = list())
}
\arguments{
\item{x}{Plan or source}

\item{what}{Character vector of assays to derive}

\item{options}{Optional named list of options}
}
\value{
Updated plan
}
\description{
Record a derivation operation in a plan
}
