% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stubs-core.R
\name{derive_eager}
\alias{derive_eager}
\title{Eagerly derive statistics and compute immediately}
\usage{
derive_eager(...)
}
\arguments{
\item{...}{Arguments passed to derive(), then compute()}
}
\value{
A realized GDS object
}
\description{
Eagerly derive statistics and compute immediately
}
