% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog.R
\name{.catalog_validation_report}
\alias{.catalog_validation_report}
\title{Create validation report object}
\usage{
.catalog_validation_report(catalog, issues)
}
\description{
Create validation report object
}
\keyword{internal}
