% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapter-nftab.R
\name{.nftab_build_space}
\alias{.nftab_build_space}
\title{Map neurotabs support/logical schema to a GDS space object}
\usage{
.nftab_build_space(logical_schema, tab, feature_name = NULL)
}
\description{
Map neurotabs support/logical schema to a GDS space object
}
\keyword{internal}
