% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapter-nftab.R
\name{.nftab_extract_col_data}
\alias{.nftab_extract_col_data}
\title{Extract subject-level covariates from the design table}
\usage{
.nftab_extract_col_data(design, subjects, subject_col)
}
\description{
Extract subject-level covariates from the design table
}
\keyword{internal}
