% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapter-nftab.R
\name{.nftab_parcel_labels}
\alias{.nftab_parcel_labels}
\title{Extract parcel labels from support info or generate defaults}
\usage{
.nftab_parcel_labels(support, tab, n)
}
\description{
Extract parcel labels from support info or generate defaults
}
\keyword{internal}
