% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapter-nftab.R
\name{.nftab_read_feature}
\alias{.nftab_read_feature}
\title{Read and reshape a single feature into a 3D array}
\usage{
.nftab_read_feature(
  tab,
  feat_name,
  design,
  subjects,
  contrasts,
  subject_col,
  contrast_col,
  sample_block
)
}
\description{
Read and reshape a single feature into a 3D array
}
\keyword{internal}
