% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapter-nftab.R
\name{.nftab_sample_count}
\alias{.nftab_sample_count}
\title{Compute the number of spatial samples from the logical schema}
\usage{
.nftab_sample_count(logical_schema, tab, feature_name = NULL)
}
\description{
Compute the number of spatial samples from the logical schema
}
\keyword{internal}
