% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapter-nftab.R
\name{.nftab_sample_labels}
\alias{.nftab_sample_labels}
\title{Get sample-level labels from axis domains or generate defaults}
\usage{
.nftab_sample_labels(logical_schema, tab, n, feature_name = NULL)
}
\description{
Get sample-level labels from axis domains or generate defaults
}
\keyword{internal}
