% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog.R
\name{.resolve_assay_map}
\alias{.resolve_assay_map}
\title{Resolve assay map to file indices}
\usage{
.resolve_assay_map(catalog, assay_name)
}
\arguments{
\item{catalog}{An image_catalog object}

\item{assay_name}{Name of the assay to resolve}
}
\value{
Integer vector of file indices
}
\description{
Resolve assay map to file indices
}
\keyword{internal}
