% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{explain}
\alias{explain}
\title{Explain objects and plans}
\usage{
explain(x)
}
\arguments{
\item{x}{A \code{gds} or \code{gds_plan}}
}
\value{
Invisibly returns \code{x} (and prints a summary)
}
\description{
Human-readable summary of a realised GDS or a lazy \code{gds_plan}.
}
