% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plan.R
\name{gds_plan}
\alias{gds_plan}
\title{Construct a lazy GDS plan}
\usage{
gds_plan(source, nodes = list(), meta = list())
}
\arguments{
\item{source}{A \code{\link{gds_source}} object}

\item{nodes}{List of operation nodes}

\item{meta}{Metadata list (planner bookkeeping)}
}
\value{
A plan object (\code{gds_plan})
}
\description{
Construct a lazy GDS plan
}
