% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapter-registry.R
\name{get_adapter}
\alias{get_adapter}
\title{Retrieve a registered adapter}
\usage{
get_adapter(name)
}
\arguments{
\item{name}{Adapter name}
}
\value{
Adapter list
}
\description{
Retrieve a registered adapter
}
