% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map-register.R
\name{get_map_family}
\alias{get_map_family}
\title{Retrieve a registered map family by name}
\usage{
get_map_family(x, name)
}
\arguments{
\item{x}{A plan or realised GDS}

\item{name}{Map family name}
}
\value{
A \code{\link{MapFamily}} object
}
\description{
Retrieve a registered map family by name
}
