% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reducer-registry.R
\name{get_reducer}
\alias{get_reducer}
\title{Get reducer by name}
\usage{
get_reducer(name)
}
\arguments{
\item{name}{Reducer name}
}
\value{
Reducer list or NULL if not found
}
\description{
Get reducer by name
}
