% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compat.R
\name{harmonise_contrasts}
\alias{harmonise_contrasts}
\title{Harmonise contrast names in a GDS}
\usage{
harmonise_contrasts(g, map)
}
\arguments{
\item{g}{A realised GDS}

\item{map}{Named character vector mapping old -> new contrast names}
}
\value{
Updated GDS with renamed contrasts
}
\description{
Harmonise contrast names in a GDS
}
\examples{
beta <- array(0, dim = c(2,1,2), dimnames = list(NULL, "s1", c("a","b")))
var <- beta
g <- new_gds(list(beta = beta, var = var), space_sample_labels(c("x","y")), "s1", c("a","b"))
harmonise_contrasts(g, c(a = "A", b = "B"))
}
