% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map-register.R
\name{list_map_families}
\alias{list_map_families}
\title{List registered map families}
\usage{
list_map_families(x)
}
\arguments{
\item{x}{A plan or realised GDS}
}
\value{
Character vector of family names
}
\description{
List registered map families
}
