% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{map_linear}
\alias{map_linear}
\title{Create a linear map between spaces}
\usage{
map_linear(
  from_space,
  to_space,
  operator,
  by_subject = NULL,
  traits = list(orthogonal = FALSE, mass_preserving = FALSE),
  uncertainty = UncertaintyRule("independent"),
  name = NULL
)
}
\arguments{
\item{from_space}{Source space descriptor or name}

\item{to_space}{Target space descriptor or name}

\item{operator}{Mapping operator (matrix/function)}

\item{by_subject}{Optional named list of subject-specific operator descriptors/functions}

\item{traits}{List of map traits (orthogonal, mass_preserving)}

\item{uncertainty}{\code{\link{UncertaintyRule}} controlling variance propagation}

\item{name}{Optional map name}
}
\value{
Object of class \code{c("map_linear", "gds_map")}
}
\description{
Create a linear map between spaces
}
