% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verb-map-to.R
\name{map_to}
\alias{map_to}
\title{Add a space transformation to a plan}
\usage{
map_to(
  x,
  target_space,
  map,
  uncertainty = UncertaintyRule("independent"),
  combine = NULL
)
}
\arguments{
\item{x}{Plan, source, or realised GDS}

\item{target_space}{Target space descriptor or identifier}

\item{map}{Mapping object (\code{map_linear} or matrix)}

\item{uncertainty}{\code{\link{UncertaintyRule}} instance controlling variance propagation}

\item{combine}{Optional evidence combiner when effect scale absent}
}
\value{
Updated plan
}
\description{
Add a space transformation to a plan
}
