% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verb-map-to.R
\name{map_to_eager}
\alias{map_to_eager}
\title{Eagerly apply space transformation and compute immediately}
\usage{
map_to_eager(x, ...)
}
\arguments{
\item{x}{Plan, source, or realized GDS}

\item{...}{Arguments passed to map_to(), then compute()}
}
\value{
A realized GDS object
}
\description{
Eagerly apply space transformation and compute immediately
}
