% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verb-mask.R
\name{mask}
\alias{mask}
\title{Apply a mask policy lazily}
\usage{
mask(x, policy = MaskPolicy())
}
\arguments{
\item{x}{Plan, source, or realised GDS}

\item{policy}{Mask policy created by \code{\link[=MaskPolicy]{MaskPolicy()}}}
}
\value{
Updated plan
}
\description{
Apply a mask policy lazily
}
