% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verb-mask.R
\name{mask_eager}
\alias{mask_eager}
\title{Eagerly apply mask policy and compute immediately}
\usage{
mask_eager(x, ...)
}
\arguments{
\item{x}{Plan, source, or realized GDS}

\item{...}{Arguments passed to mask(), then compute()}
}
\value{
A realized GDS object
}
\description{
Eagerly apply mask policy and compute immediately
}
