% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gds-class.R
\name{metadata}
\alias{metadata}
\title{Extract metadata from a GDS object}
\usage{
metadata(x)
}
\arguments{
\item{x}{A GDS object}
}
\value{
Metadata list
}
\description{
Extract metadata from a GDS object
}
