% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interop.R
\name{model_matrix}
\alias{model_matrix}
\title{Build a design matrix from attached col_data}
\usage{
model_matrix(x, formula)
}
\arguments{
\item{x}{A \code{gds} or \code{gds_plan} (or an object coercible via \code{\link[=as_plan]{as_plan()}})}

\item{formula}{Model formula (or string convertible via \link[stats:formula]{stats::as.formula})}
}
\value{
A numeric model matrix
}
\description{
Convenience helper to construct a model matrix using subject-level
covariates already attached to a plan via \code{\link[=with_col_data]{with_col_data()}} or stored in a
realised GDS.
}
