% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plan.R
\name{op_align_to_group}
\alias{op_align_to_group}
\title{Create an align-to-group operation node}
\usage{
op_align_to_group(family = NULL, family_name = NULL)
}
\arguments{
\item{family}{MapFamily object for alignment}

\item{family_name}{Name of registered map family}
}
\value{
Operation node list
}
\description{
Create an align-to-group operation node
}
