% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plan.R
\name{op_derive}
\alias{op_derive}
\title{Create a derive operation node}
\usage{
op_derive(what, options = list())
}
\arguments{
\item{what}{Character vector of statistics to derive (e.g., "t", "z", "p")}

\item{options}{Optional list of derivation options}
}
\value{
Operation node list
}
\description{
Create a derive operation node
}
