% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plan.R
\name{op_mask_policy}
\alias{op_mask_policy}
\title{Create a mask policy operation node}
\usage{
op_mask_policy(policy)
}
\arguments{
\item{policy}{MaskPolicy object}
}
\value{
Operation node list
}
\description{
Create a mask policy operation node
}
