% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plan.R
\name{op_subset_axis}
\alias{op_subset_axis}
\title{Create a subset operation node}
\usage{
op_subset_axis(sample = NULL, subject = NULL, contrast = NULL)
}
\arguments{
\item{sample}{Sample indices or labels}

\item{subject}{Subject indices or IDs}

\item{contrast}{Contrast indices or names}
}
\value{
Operation node list
}
\description{
Create a subset operation node
}
