% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plan-alias.R
\name{plan}
\alias{plan}
\title{Coerce to a lazy plan (alias)}
\usage{
plan(x)
}
\arguments{
\item{x}{A \code{gds}, \code{gds_source}, or \code{gds_plan}}
}
\value{
A \code{gds_plan}
}
\description{
Alias for \code{\link[=as_plan]{as_plan()}} to improve ergonomics when converting realized GDS
objects or sources into a \code{gds_plan}.
}
