% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog.R
\name{print.catalog_validation_report}
\alias{print.catalog_validation_report}
\title{Print validation report}
\usage{
\method{print}{catalog_validation_report}(x, ...)
}
\arguments{
\item{x}{A catalog_validation_report object.}

\item{...}{Additional arguments (ignored).}
}
\value{
Invisibly returns the report.
}
\description{
Print validation report
}
