% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog.R
\name{print.image_catalog}
\alias{print.image_catalog}
\title{Print method for image_catalog}
\usage{
\method{print}{image_catalog}(x, ...)
}
\arguments{
\item{x}{An image_catalog object.}

\item{...}{Additional arguments (ignored).}
}
\value{
Invisibly returns the catalog.
}
\description{
Print method for image_catalog
}
