% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propagate-variance.R
\name{propagate_variance_independent}
\alias{propagate_variance_independent}
\title{Propagate variance assuming independent samples}
\usage{
propagate_variance_independent(M, beta, var)
}
\arguments{
\item{M}{Mapping matrix \code{[n_target x n_source]}}

\item{beta}{Effect array \code{[n_source x n_subject x n_contrast]}}

\item{var}{Variance array (same dims as beta)}
}
\value{
List with \code{beta} and \code{var} arrays in target space
}
\description{
Propagate variance assuming independent samples
}
\keyword{internal}
