% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verb-reduce.R
\name{reduce_eager}
\alias{reduce_eager}
\title{Eagerly reduce across subjects and compute immediately}
\usage{
reduce_eager(x, ...)
}
\arguments{
\item{x}{Plan, source, or realized GDS}

\item{...}{Arguments passed to reduce(), then compute()}
}
\value{
A realized GDS object
}
\description{
Eagerly reduce across subjects and compute immediately
}
