% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compat.R
\name{relabel_subjects}
\alias{relabel_subjects}
\title{Relabel subjects in a GDS}
\usage{
relabel_subjects(g, mapping)
}
\arguments{
\item{g}{A realised GDS}

\item{mapping}{Named character vector mapping old -> new subject ids}
}
\value{
Updated GDS with relabeled subjects
}
\description{
Relabel subjects in a GDS
}
\examples{
beta <- array(0, dim = c(2,2,1), dimnames = list(NULL, c("u1","u2"), "c1"))
var <- beta
g <- new_gds(list(beta = beta, var = var), space_sample_labels(c("x","y")), c("u1","u2"), "c1")
relabel_subjects(g, c(u1 = "s1", u2 = "s2"))
}
