% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gds-class.R
\name{row_data}
\alias{row_data}
\title{Extract row (sample) metadata from a GDS object}
\usage{
row_data(x)
}
\arguments{
\item{x}{A GDS object}
}
\value{
Data frame with sample-level metadata
}
\description{
Extract row (sample) metadata from a GDS object
}
