% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gds-class.R
\name{sample_labels}
\alias{sample_labels}
\title{Extract sample labels from a GDS object}
\usage{
sample_labels(x)
}
\arguments{
\item{x}{A GDS object or plan}
}
\value{
Character vector of sample labels
}
\description{
Extract sample labels from a GDS object
}
