% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gds-class.R
\name{space}
\alias{space}
\title{Extract space descriptor from a GDS object}
\usage{
space(x)
}
\arguments{
\item{x}{A GDS object}
}
\value{
A space object
}
\description{
Extract space descriptor from a GDS object
}
