% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/space.R
\name{space_basis}
\alias{space_basis}
\title{Create a latent basis space descriptor}
\usage{
space_basis(k, basis_name = NULL, projector = NULL, voxel_space = NULL)
}
\arguments{
\item{k}{Number of components}

\item{basis_name}{Optional name for the basis (e.g., "ICA")}

\item{projector}{Optional matrix/function for voxel projection}

\item{voxel_space}{Optional reference voxel space descriptor}
}
\value{
Space object
}
\description{
Create a latent basis space descriptor
}
