% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/space-utils.R
\name{space_from_nifti}
\alias{space_from_nifti}
\title{Create a voxel space from a NIfTI file}
\usage{
space_from_nifti(path, mask = NULL)
}
\arguments{
\item{path}{Path to a NIfTI file (.nii or .nii.gz)}

\item{mask}{Optional path to a mask NIfTI or a neuroim2::NeuroVol}
}
\value{
A \code{space_voxel} object
}
\description{
Create a voxel space from a NIfTI file
}
\examples{
# if (requireNamespace("neuroim2", quietly = TRUE)) {
#   sp <- space_from_nifti(system.file("extdata", "tiny.nii", package = "fmrigds"))
# }
}
