% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/space.R
\name{space_parcels}
\alias{space_parcels}
\title{Create a parcels/ROI space descriptor}
\usage{
space_parcels(labels, lookup = NULL, membership = NULL)
}
\arguments{
\item{labels}{Character vector of parcel labels}

\item{lookup}{Optional lookup information (data frame or list)}

\item{membership}{Optional mapping (list or sparse matrix) from parcels to voxels}
}
\value{
Space object
}
\description{
Create a parcels/ROI space descriptor
}
