% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examination-staging.R
\name{space_portability}
\alias{space_portability}
\title{Classify the portability of a GDS space}
\usage{
space_portability(space)
}
\arguments{
\item{space}{A GDS space object.}
}
\value{
One of \code{"portable"}, \code{"portable_with_external_reference"}, or
\code{"nonportable_function"}.
}
\description{
Classify the portability of a GDS space
}
