% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/space.R
\name{space_sample_labels}
\alias{space_sample_labels}
\title{Create a simple label space for tabular samples}
\usage{
space_sample_labels(labels)
}
\arguments{
\item{labels}{Character vector of sample labels}
}
\value{
Space object of type "sample_labels"
}
\description{
Create a simple label space for tabular samples
}
