% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dims-class.R
\name{[.gds_dims}
\alias{[.gds_dims}
\title{Subset method strips names to avoid attribute sensitivity in tests}
\usage{
\method{[}{gds_dims}(x, i, ...)
}
\description{
Subset method strips names to avoid attribute sensitivity in tests
}
\keyword{internal}
