% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gds-class.R
\name{subjects}
\alias{subjects}
\title{Extract subject identifiers from a GDS object}
\usage{
subjects(x)
}
\arguments{
\item{x}{A GDS object}
}
\value{
Character vector of subject IDs
}
\description{
Extract subject identifiers from a GDS object
}
