% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stubs-core.R
\name{subset_eager}
\alias{subset_eager}
\title{Eagerly subset a GDS and compute immediately}
\usage{
subset_eager(...)
}
\arguments{
\item{...}{Arguments passed to subset(), then compute()}
}
\value{
A realized GDS object
}
\description{
Eagerly subset a GDS and compute immediately
}
