% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog.R
\name{summary.image_catalog}
\alias{summary.image_catalog}
\title{Summary method for image_catalog}
\usage{
\method{summary}{image_catalog}(object, ...)
}
\arguments{
\item{object}{An image_catalog object.}

\item{...}{Additional arguments (ignored).}
}
\value{
Invisibly returns the catalog.
}
\description{
Summary method for image_catalog
}
