% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posthoc-registry.R
\name{unregister_posthoc}
\alias{unregister_posthoc}
\title{Unregister a post-hoc method}
\usage{
unregister_posthoc(name)
}
\arguments{
\item{name}{Method identifier to remove}
}
\value{
Invisibly, TRUE if removed, FALSE if not found
}
\description{
Unregister a post-hoc method
}
