% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{validate}
\alias{validate}
\title{Validate a GDS or plan}
\usage{
validate(x, ...)
}
\arguments{
\item{x}{A \code{gds}, \code{gds_plan}, or \code{image_catalog}}

\item{...}{Additional arguments passed to methods}
}
\value{
TRUE if valid; otherwise errors with a descriptive message
}
\description{
Validate a GDS or plan
}
