% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reducer-options.R
\name{validate_reducer_options}
\alias{validate_reducer_options}
\title{Validate reducer options against schema}
\usage{
validate_reducer_options(schema, opts)
}
\description{
Validate reducer options against schema
}
\keyword{internal}
