% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gds-class.R
\name{contrasts}
\alias{contrasts}
\title{Extract contrast identifiers from a GDS object}
\usage{
contrasts(x)
}
\arguments{
\item{x}{A GDS object}
}
\value{
Character vector of contrast names
}
\description{
Extract contrast identifiers from a GDS object
}
