% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plan-serialization.R
\name{load_plan}
\alias{load_plan}
\title{Load a plan from JSON}
\usage{
load_plan(file)
}
\arguments{
\item{file}{JSON file produced by \link{save_plan}}
}
\value{
A \code{gds_plan}
}
\description{
Load a plan from JSON
}
