% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verb-mask.R
\name{MaskPolicy}
\alias{MaskPolicy}
\title{Define a mask policy}
\usage{
MaskPolicy(
  scope = c("group", "subject"),
  rule = c("intersection", "union", "threshold", "custom"),
  threshold = 0.95,
  custom = NULL
)
}
\arguments{
\item{scope}{Scope of mask computation ("group" or "subject")}

\item{rule}{Mask rule ("intersection", "union", "threshold", "custom")}

\item{threshold}{Fraction for threshold rule}

\item{custom}{Custom function returning logical mask}
}
\value{
A mask policy object
}
\description{
Define a mask policy
}
