% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gds-class.R
\name{provenance_node}
\alias{provenance_node}
\title{Create a provenance node}
\usage{
provenance_node(
  op_name,
  params,
  inputs = list(),
  timestamp = Sys.time(),
  software = list(package = "gdsfmri", version = .pkg_version()),
  hash = NULL
)
}
\arguments{
\item{op_name}{Operation name}

\item{params}{Named list of parameters}

\item{inputs}{Vector/list of parent hashes/ids}

\item{timestamp}{Timestamp of operation}

\item{software}{Software metadata}

\item{hash}{Optional pre-computed hash}
}
\value{
A provenance node list
}
\description{
Create a provenance node
}
