% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog.R
\name{read_catalog}
\alias{read_catalog}
\title{Read catalog from JSON file}
\usage{
read_catalog(file)
}
\arguments{
\item{file}{Path to catalog JSON file.}
}
\value{
An image_catalog object.
}
\description{
Load an image catalog previously saved with \code{\link[=write_catalog]{write_catalog()}}.
}
\examples{
\dontrun{
catalog <- read_catalog("my_study_manifest.json")
}
}
