% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map-register.R
\name{register_map}
\alias{register_map}
\title{Register a map family on a plan or realised GDS}
\usage{
register_map(x, family, overwrite = FALSE)
}
\arguments{
\item{x}{A \code{\link{gds_plan}} or realised \code{\link{gds}}}

\item{family}{A \code{\link{MapFamily}} object}

\item{overwrite}{Whether to overwrite an existing family with the same name}
}
\value{
Updated object (\code{gds_plan} or \code{gds})
}
\description{
Register a map family on a plan or realised GDS
}
