% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plan-serialization.R
\name{save_plan}
\alias{save_plan}
\title{Save a plan to JSON}
\usage{
save_plan(plan, file)
}
\arguments{
\item{plan}{Plan or object coercible via \code{\link[=as_plan]{as_plan()}}}

\item{file}{Path to JSON file}
}
\description{
Save a plan to JSON
}
