% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/space.R
\name{space_surface}
\alias{space_surface}
\title{Create a surface space descriptor}
\usage{
space_surface(vertices, faces, hemi, template_id = NULL)
}
\arguments{
\item{vertices}{Matrix (n x 3) of vertex coordinates}

\item{faces}{Integer matrix (m x 3) of triangle indices (1-based)}

\item{hemi}{Hemisphere identifier ("L", "R", "LR")}

\item{template_id}{Optional template identifier}
}
\value{
Space object
}
\description{
Create a surface space descriptor
}
