% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights.R
\name{use_weight}
\alias{use_weight}
\title{Convenience helper to use a stored weight array for reduction}
\usage{
use_weight(g, name)
}
\arguments{
\item{g}{A realised GDS containing the weight assay}

\item{name}{Assay name to use as custom weights}
}
\value{
A list(list) suitable to splice into reduce():
list(weights = "custom", options = list(custom_weights = \if{html}{\out{<array>}}))
}
\description{
Convenience helper to use a stored weight array for reduction
}
\examples{
# opts <- use_weight(g2, "w_custom")
}
